+++ CHANGES.md
@@ -1,3 +1,4 @@
+- Reset client_caps on every plugin (de)activation
 - Replaced occurences of `is_plugin_active` by `class_exists`
 - Added reset link to plugins page
 - Added function remove_client_caps